Rename data_free.py to svdquant.py
The module implements the SVDQuant math (smoothing, low-rank split,
quantization, kernel packing) as opposed to utils.py's kernel runtime;
name it after the algorithm. Data-free stays in the function names,
where it describes the mode.
